+++ resources/js/components/order/OrderListTable.vue
@@ -23,57 +23,7 @@
 
     import {mapGetters, mapActions} from 'vuex';
     import InvoiceModalView from "../invoice/InvoiceModalView";
-
-    Vue.component('order-status-badge', {
-        template: `<div class="btn-group">
-                        <span  class="label  dropdown-toggle"
-                            :class="{'bg-info':row.order_status == 1, 'bg-danger':row.order_status == 2, 'bg-warning':row.order_status == 3, 'bg-primary':row.order_status == 4, 'bg-indigo-400':row.order_status == 5, 'bg-teal':row.order_status == 6 }"
-                            data-toggle="dropdown" aria-expanded="false">
-                            {{ row.status_label }}
-                            <span v-if="row.order_status != 2  && row.order_status != 6" class="caret"></span>
-                        </span>
-                        <div v-if="row.order_status != 2 && row.order_status != 6" >
-                            <ul v-if="statusList" class="dropdown-menu dropdown-menu-right">
-                              <li v-for="(status, index) in statusList" v-if="row.order_status != index">
-                                  <a href="#" @click.prevent="changeOrderStatus(row.order_no, index)">
-                                    <span
-                                    class="status-mark position-left"
-                                    :class="{'bg-info':index == 1, 'bg-danger':index == 2, 'bg-warning':index == 3, 'bg-primary':index == 4, 'bg-indigo-400':index == 5, 'bg-teal':index == 6 }"
-                                    ></span> {{ status }}
-                                  </a>
-                              </li>
-                            </ul>
-                        </div>
-                    </div>`,
-        props: ['row'],
-        methods:{
-            ...mapActions([
-                'orderStatusChange'
-            ]),
-            changeOrderStatus(orderNo, status){
-                //TODO use nice Global Modal For Confirm Action
-
-                let conf = confirm('Are You Sure Want To change Order Status.?');
-                if(!conf){
-                    return false;
-                }
-                this.orderStatusChange({order_no:orderNo, status:status})
-                    .then(response=>{
-                        if(typeof response.code !== "undefined" && response.code === 200){
-                            Notify.success(response.message);
-                        }else{
-                            Notify.error(response.message);
-                        }
-                    })
-            }
-        },
-        computed:{
-            ...mapGetters([
-                'statusList'
-            ])
-        }
-
-    });
+    import OrderStatusBadge from "../helper/table/OrderStatusBadge";
 
     Vue.component('action-btn', {
         template: `<ul class="icons-list">
@@ -112,10 +62,21 @@
         },
     });
 
+    Vue.component('product-list', {
+        template: `<div>
+                       <ul v-if="row.order_items">
+                            <li v-for="(oItem,index) in row.order_items">{{ oItem.product_name }} - ({{ oItem.qty }})</li>
+                        </ul>
+                    </div>`,
+        props: ['row']
+    });
 
     export default {
         name: "OrderListTable",
-        components: {InvoiceModalView},
+        components: {
+            InvoiceModalView,
+            'order-status-badge':OrderStatusBadge,
+        },
         props:['reqData'],
         data(){
             return {
@@ -128,6 +89,7 @@
                     { label: 'Order No', field: 'order_no',  },
                     { label: 'Order Date', field: 'order_date', filterable: true, sortable:true },
                     { label: 'Buyer', field: 'buyer.user.full_name' , filterable: true, sortable:true },
+                    { label: 'Products - (Qty)', component:'product-list', filterable: true, sortable:false },
                     { label: 'Quantity', field: 'total_qty', align: 'right',  },
                     { label: 'Subtotal', field: 'sub_total', align: 'right', sortable: true },
                     { label: 'Total', field: 'total', align: 'right', sortable: true },
